chore: fix review comment, add interface.py content in user-client-state
