+++ lms/djangoapps/courseware/tests/test_user_state_client.py
@@ -12,9 +12,8 @@ from collections import defaultdict
 from django.db import connections
 
 from common.djangoapps.student.tests.factories import UserFactory
-from lms.djangoapps.courseware.user_state_client import DjangoXBlockUserStateClient
+from lms.djangoapps.courseware.user_state_client import DjangoXBlockUserStateClient, XBlockUserStateClient, XBlockUserState
 from xmodule.modulestore.tests.django_utils import ModuleStoreTestCase  # lint-amnesty, pylint: disable=wrong-import-order
-from lms.djangoapps.courseware.interface import XBlockUserStateClient, XBlockUserState
 
 
 class _UserStateClientTestUtils(TestCase):
